requirements.txt
# This requirements file lists all third-party dependencies for this project.
#
# Run 'pip install -r requirements.txt -t lib/' to install these dependencies
# in `lib/` subdirectory.
#
# Note: The `lib` directory is added to `sys.path` by `appengine_config.py`.
python-dateutil==2.4.2
pytz==2015.7
Flask==0.10.1
Flask-Login==0.3.2
Flask-WTF==0.12
Flask-Assets==0.10
oauth2client==1.4.2
git+git://github.com/google/google-api-python-client@master
